+++ tools/idea-plugin/CHANGELOG.md
@@ -13,14 +13,15 @@
 
 ### Changed
 
-- [IconPack] Don’t override the `.kt` file object if the content hasn’t changed
+- [IconPack] Don't override the `.kt` file object if the content hasn't changed
+- [IconPack] Generated ImageVector files are now appear immediately in the project tree without manual refresh
 
 ### Fixed
 
 - [Gutter] Fixed rendering of icons with gradients and Compose colors
 - [Gutter] Fixed rendering of icons with fully qualified imports
 - [IconPack] Fix incorrect initial state of `packageName` field in new pack mode when destination folder is selected
-- [IconPack] The IDE didn’t automatically display the new icon pack object file
+- [IconPack] The IDE didn't automatically display the new icon pack object file
 
 ## 1.5.0 - 2026-04-17